Organize Sass files and folders, remove files/code not needed, scroll bar on trending NFTs page
