+++ src/css/main.css
@@ -1,18 +1,30 @@
 body {
-  background-color: white;
-}
-
-body {
   background-color: #cbe6ef;
 }
 
-.navbar .container .navbar-brand {
-  font-size: 35px;
-  margin: 8px 0;
-  font-family: "Franklin Gothic Medium", "Arial Narrow", Arial, sans-serif;
+.video-responsive {
+  overflow: hidden;
+  padding-bottom: 56.25%;
+  position: relative;
+  height: 0;
 }
-.navbar .container .navbar-brand .nav .nav-link {
-  font-size: 1.6rem;
+.video-responsive iframe {
+  left: 0;
+  top: 0;
+  height: 100%;
+  width: 100%;
+  position: absolute;
+}
+
+.news-body {
+  display: flex;
+}
+.news-body .daily-news {
+  width: 70%;
+  margin-left: 3rem;
+}
+.news-body .nft-list {
+  width: 30%;
 }
 
 .about {
@@ -132,6 +144,69 @@ body {
   transition: background-color 0.5s ease-out;
 }
 
+.list-title-container {
+  display: flex;
+  height: 110px;
+  align-items: center;
+  justify-content: center;
+}
+.list-title-container .nft-list-title {
+  text-align: center;
+  margin: 0;
+  padding: 0;
+}
+
+.nft-list {
+  height: 89vh;
+  overflow-y: auto;
+}
+
+.nft-card {
+  border-radius: 1.5rem;
+  transition: all 0.2s ease-in;
+}
+
+.nft-card:hover {
+  transform: scale(1.03);
+}
+
+form {
+  max-width: 50%;
+  height: 100%;
+  padding: 0.5rem;
+  margin-bottom: 1rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+}
+form option {
+  color: black;
+}
+form label {
+  margin-bottom: 1rem;
+}
+
+.page-head-container {
+  display: flex;
+  justify-content: space-evenly;
+  height: 110px;
+}
+.page-head-container h1 {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.navbar .container .navbar-brand {
+  font-size: 35px;
+  margin: 8px 0;
+  font-family: "Franklin Gothic Medium", "Arial Narrow", Arial, sans-serif;
+}
+.navbar .container .navbar-brand .nav .nav-link {
+  font-size: 1.6rem;
+}
+
 .header {
   flex-grow: 1;
   display: flex;
@@ -165,79 +240,4 @@ body {
   color: #f2f2f2;
   font-size: 20px;
   font-family: "Trebuchet MS", "Lucida Sans Unicode", "Lucida Grande", "Lucida Sans", Arial, sans-serif;
-}
-
-.list-title-container {
-  display: flex;
-  height: 110px;
-  align-items: center;
-  justify-content: center;
-}
-.list-title-container .nft-list-title {
-  text-align: center;
-  margin: 0;
-  padding: 0;
-}
-
-.nft-card {
-  border-radius: 1.5rem;
-  transition: all 0.2s ease-in;
-}
-
-.nft-card:hover {
-  transform: scale(1.03);
-}
-
-.page-body {
-  display: flex;
-  background-color: #cbe6ef;
-}
-.page-body .daily-news {
-  width: 60%;
-}
-.page-body .nft-list {
-  width: 40%;
-}
-
-form {
-  max-width: 50%;
-  height: 100%;
-  padding: 0.5rem;
-  margin-bottom: 1rem;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-form option {
-  color: black;
-}
-form label {
-  margin-bottom: 1rem;
-}
-
-.page-head-container {
-  display: flex;
-  justify-content: space-evenly;
-  background-color: #cbe6ef;
-  height: 110px;
-}
-.page-head-container h1 {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.video-responsive {
-  overflow: hidden;
-  padding-bottom: 56.25%;
-  position: relative;
-  height: 0;
-}
-.video-responsive iframe {
-  left: 0;
-  top: 0;
-  height: 100%;
-  width: 100%;
-  position: absolute;
 }
