+++ src/sass/base/_base.scss
@@ -1,3 +1,11 @@
+// VARIABLES
+$backgroundColor: rgb(203,230,239);
+
+// TYPOGRAPHY
+
+
+// OVERALL
+
 body {
-    background-color: white;
+    background-color: $backgroundColor;
 }
