--- src/sass/components/_page-body.scss
@@ -1,14 +1,0 @@
-.page-body{
-    display: flex;
-    background-color: rgb(203,230,239);
-    
-    .daily-news{
-        width: 60%;
-    }
-
-    .nft-list{
-        width: 40%;
-        
-    }
-}
-
