+++ src/sass/components/_page-header.scss
@@ -1,8 +1,6 @@
-$backgroundColor: rgb(203,230,239);
 .page-head-container{
     display: flex;
     justify-content: space-evenly;
-    background-color: $backgroundColor;
     height: 110px;
 
     h1{
