+++ src/sass/pages/_news.scss
@@ -1,0 +1,13 @@
+.news-body{
+    display: flex;
+
+    .daily-news{
+        width: 70%;
+        margin-left: 3rem;
+    }
+
+    .nft-list{
+        width: 30%;        
+    }
+}
+
