+++ src/sass/pages/_resources.scss
