--- src/sass/pages/_trendingNFTs.scss
@@ -1,3 +1,0 @@
-body {
-    background-color: rgb(203,230,239);
-}
